+++ Intel-Systems/common/isbc201.c
@@ -206,7 +206,7 @@
 
 /* external globals */
 
-extern int32    PCX;
+extern uint16    PCX;
 
 /* external function prototypes */
