+++ Intel-Systems/common/isbc206.c
@@ -174,7 +174,7 @@
 
 /* external globals */
 
-extern int32    PCX;
+extern uint16    PCX;
 
 /* external function prototypes */
